[Android] Decouple runtime initialization and entry point execution for Android sample (#111742)
By decoupling the initialization of runtime and execution of entry point, we can run the runtime initialization synchronously, forcing the onCreate callback to report the Displayed <name> for user 0: <time> after the runtime is fully initialized.

The execution of the entry point is still performed asynchronously to prevent the freeze of main thread for strenuous executables.

---------

Co-authored-by: Ivan Povazan <55002338+ivanpovazan@users.noreply.github.com>