Rework Unix socket option tracking (#111676)
This PR changes how socket option tracking and state transfer is implemented to support multi-connect on Unix. Instead of calling property setters and getters, the new logic uses `GetSocketOption` / `SetSocketOption` directly and extends the set of tracked options. The property logic did two things: validations and and the selection of the the necessary `SocketOptionLevel`. Regarding the former: when transferring the state, validation is unnecessary, since it already happened in the property setter of the original socket. Regarding the latter: `SocketOptionLevel` is now implied by `TrackableSocketOptions`. To make sure there is no mistake in the mapping, I generated the necessary mapping logic programatically based on option names/values taken from Windows headers: https://gist.github.com/antonfirsov/2cbfc37e665ad840ed7734994948c29a

The new implementation now tracks more options, however options which are specific to datagram/udp sockets were removed from the set of tracked options, since such sockets will never be multi-connected in practice. If we think we need to add some of them back, it is easy to do so since it is easy to extend the new tracking logic to other known options of `int` values.

Additionally, the PR relaxes the check done in `ValidateForMultiConnect`:  if the socket handle's state is still transferrable (`ExposedHandleOrUntrackedConfiguration == false`), we no longer throw from multi-endpoint API-s at the beginning of the call. If the first connect attempt fails and the handle is exposed, we terminate with the exception of the first failure, ignoring the remaining endpoints.