Use enum for frametype not v table (#112166)
This change refactors the frame identification mechanism in the runtime by replacing the use of the VTable with an enum FrameIdentifier. The VTable-based approach more suited to C++ code, but presented challenges for assembly code and out-of-process inspections. By using FrameIdentifier, polymorphic dispatch is now handled via if-statements. This change also eliminates the need for GS cookies on Frame objects, as it is no longer necessary to protect the VTable.

Key changes include:

Renaming FrameType to FrameIdentifier
Introducing a GetFrameIdentifier method
This update streamlines frame management while improving performance and flexibility for various platforms.