JIT: generalize ParseArrayAddress a bit more
Now that we are stack allocating arrays, the array reference may be either TYP_BYREF
or TYP_I_IMPL, so try and recognize those cases when possible. In particular for OSR
methods the array allocation may "disappear" once we redirect flow to the OSR entry,
so relying on VN is insufficient.

Fixes #112429 (benchmark regression where we're measuring OSR code)