JIT: enhance escape analysis to understand `Span<T>` capture
Implement a very simplistic "field sensitive" analysis for `Span<T>`
where we model the span as simply its byref field.

If the span is only consumed locally, and the array does not otherwise
escape, then the array does not escape.

This is a subset of #112543 that does not try and reason interprocedurally.

Contributes to #104936 / #108913