Run TypeIntrinsics tests with native AOT (#112477)
I don't think the `CreateDynamic` flavors test anything interesting from codegen perspective. The `IsValueType` and `IsValueTypeObj` methods are going to be called using reflection, but nothing interesting will happen in the method body (it's a very regular method body). All the `dynamic` keyword is doing there is that a ton of reflection is going to happen before we execute the regular method body.