Fix win-arm64 crossdac build
https://github.com/dotnet/runtime/pull/112256 added the ability to cross compile android on windows. This
regressed the win-arm64 crossdac build because it stopped passing CMAKE_TARGET_ARCH and CMAKE_TARGET_OS for
cross component builds where the host and target os's match.

This change makes sure the cmake variables are passed in all cases except when the hostos isn't passed in or is
windows (host == target).