Fix new EH hang on DebugBreak
The new exception handling doesn't work well with the DebugBreak in some
cases. E.g. when it is invoked from FATAL_GC_ERROR. The new EH attempts
to handle the STATUS_BREAKPOINT stemming from the DebugBreak, allocate a
managed exception object and hangs since it cannot do that when the GC
is running.

The cause is a missing check for the breakpoint exception in the
ProcessCLRExceptionNew that is present in the old ProcessCLRException.
To fix it, I've copied that code to the ProcessCLRExceptionNew.

Close #112599