[WIP] Remove android_coreclr_initialize and modify the existing one
Signature is altered on Android, by adding a pointer to the host contract
structure.

The host contract structure also gains a new pointer, `external_assembly_probe`
(proposed in https://github.com/dotnet/runtime/issues/112706) which is used, if
present, when `bundle_probe` is `nullptr` or if it returns `false` when called.

Either `bundle_probe` or `external_assembly_probe` must be set in the contract