Bump the minimum ICU version to 60 (#112713)
* Bump the minimum ICU version to 60

As ICU approaches version 80, we need to consider what to do about our ICU version ranges. Our system currently supports a min version of 50 and a max of "min + 30", so 80.

ICU60 is the version that shipped with Ubuntu 18.04, which is the distro that we base our libc version on. For Alpine, the version of ICU in the distro we base our libc on (3.17) ships with ICU 72.

I recommend that we bump our min to ICU60, which will by our current rules bump our max to ICU90, which will give us more runway.

I'd recommend holding off on merging this until https://github.com/dotnet/runtime/pull/112671 is merged, as that PR changes our deps packages to better describe scenarios where we actually ship packages today.

* Remove old optional definition