Hotfix interpreter crash
The fake unwind info generation in the interpreter was broken and caused
write after an end of the allocated space.

I am working on a change to add a interpreter specific code manager that
gets rid of the need for fake unwind info creation, but to prevent CI
failures for now, I have created this intermediate hotfix.