Add DeriveKeyFromSessionKey API for auth
Added the DeriveKeyFromSessionKey function to NegotiateAuthentication
which can be used to retrieve the session key negotiated between the
client and server in a negotiated authentication context. This can be
used by protocols that do not use the builtin wrapping mechanisms but
instead derive their own authentication method from the session key.

Fix #111099