SPMI: Add fallback for `notifyInstructionSetUsage` (#112925)
I noticed that `notifyInstructionSetUsage` misses are pretty common,
particularly because we query for instruction sets for some very basic
instructions (e.g. `GT_RSZ/`GT_RSH`). So changes that introduce new
kinds of nodes now seem to be much more likely to miss.

The best solution would probably be to add another set of ISAs that is
passed to the JIT so that `notifyInstructionSetUsage` can be purely a
"result" type of JIT-EE call, but this change at least gets us back to
what we had before with an improvement to when we actually know the
answer.