JIT: Avoid unnecessary write backs during physical promotion
Physical promotion sometimes needs to insert writebacks to restore the
struct local from a promoted replacement field. This is required if the
field is more up to date and we end up in a situation where we need to
represent something using the struct local.

Before this change the handling of this was quite simplistic. We assumed
that the replacement was more up to date than the struct local at the
beginning of every basic block.
This PR instead adds flow sensitive tracking to propagate this
information across basic blocks. We now only consider the replacement
local to be more up to date than the struct local if it was actually
written in any of the predecessors.