Differentiate `MethodTable`s that are/aren't visible to reflection (#112782)
Fixes #112029.

We use constructed/necessary `MethodTable`s to distinguish between MTs that are visible to user code vs those that aren't. We use unconstructed MTs for things like casts or `typeof` comparisons, and some of our data structures. We use constructed MTs for everything else. Necessary MT can skip generating some of the data such as GCDesc, vtable, or metadata.

The constructed/necessary `MethodTable`s get reconciled at object writing - if anything needed a constructed one, all references to necessary MTs gets rewritten to point to the constructed one (we never have two MTs for a single type).

Before this PR, we didn't have a "constructed" MT for uninstantiated generic type definitions. That's because those don't have GCDesc or vtable. But they still have metadata, so there is a possible saving.

This introduces a "constructed"/necessary distinction for generic type definition MTs as well. The difference is in the availability of metadata.

It fixes cases like in the referenced bug where using a generic type definition in a typeof comparison (that we normally optimize) doesn't lead to the metadata being optimized away.