[android] Make native debugging test apps a little easier
To native debug the runtime in android studio, all runtime native lib symbols need to end in .so as opposed to .dbg. We have yet to find the way for Android Studio to automatically pick this up and so you have to manually associate libs to symbol lib in the UI. The UI filters based on .so and that is the reason for the rename.

This change adds a target in the test build where it'll copy .dbg files into the apk lib/arch folder and rename the extension to .so. When you build the test, you can open the apk in Android Studio and you won't have to hunt very far to associate the symbol libraries.