Remove ValidationContext.MemberType
This property is unused, and is only ever set. It has an UnconditionalSuppressMessage suppression, which makes it hard to tell if there is a trimming issue here or not.

Since this code isn't used, it is better to just remove it to make the trim analysis of this class easier.