Fix exception propagation from managed to native on Unix (#112957)
* Fix exception propagation from managed to native on Unix

The new exception handling on Unix doesn't prevent propagation of exceptions
from managed to external native code like the old one did. Instead of
reporting an unhandled exception, it rethrows the PAL_SEHException that
later ends up being reported as C++ unhandled exception.

This change fixes that behavior by enabling exception propagation from managed
to native code only for the case when the native caller is CallDescrWorkerInternal.

* Fix condition polarity

When doing final rebasing on main where the IsCallDescrWorkerInternalReturnAddress
was added in the meanwhile, I've acidentally caused the condition to be reverse
for the added check. That caused the CI test failures.

* Fix case of exception in filter

The exception in filter was incorrectly rejected as coming from external native code