Establish runtime-diagnostics pipeline
The idea behind this pipeline is to pull in dotnet/diagnostics in order to supply live runtimes to diagnostics test runs. Although it will run on a schedule, the primary reason to have this is for PR validation when making changes we know could impact the diagnostics repo.

Note that this change is bare bones in order to establish the pipeline in AzDo.