ILLink : Fix instantiation tracking issue with icall/pinvoke parameters
Along the lines of https://github.com/dotnet/runtime/pull/113434.  The existing `ProcessInteropMethod` has a bug where the return type and/or out parameter types will not be marked as instantiated.  In this case, the bug happens when the return type or parameter types do not have a `.ctor()`.  This is because `ProcessInteropMethod` will call `MarkDefaultConstructor` on the type which will cause the type to be marked instantiated if the `.ctor()` exists.