Add ms:format-date and ms:format-time test cases
Test cases for regression bug #93189 (difference between 4.52 and 6.0)

NOTE: The baseline\bug9389.xml reflects the _current_ (broken) behavior so these tests will currently be PASSING. When the bug is resolved, this test data needs to be updated to reflect what _actually_ should happen.

Fix problem with writing out the diff.xml file when test fails
You have to open the file in Create mode with Write access.

Turns out that the separator between the minutes and the AM/PM is not a space, it's actually a n U+202F NARROW NO-BREAK SPACE so update the expected value.

Add code to normalize spaces in the XmlDiff. Since it's hard to see different kinds of spaces, add the option to treat them all as normal spaces.