Try improving reliability of TestLongProcessIsWorking
Pull request #111850 improved reliability of this test for a bit but after we upgraded Helix machines to the new macOS, this test started very reliably fail with native AOT. My theory is that copying the binary around makes macOS treat the binary as a new thing that needs to be checked for signing/malware/etc and this takes a long time. So stop copying the binary.